+++ include/parallel.h
@@ -11,10 +11,10 @@ namespace Parallel {
 void tile_propagate_layer(Layer *l, int neuron_start, int neuron_end);
 
 void tile_layer_train(Layer *l, Layer *pl, int neuron_start, int neuron_end,
-                      double learning_rate);
+                      float learning_rate);
 
 void tile_layer_update(Layer *l, int neuron_start, int neuron_end,
-                       double learning_rate, double momentum);
+                       float learning_rate, float momentum);
 }
 
 #endif
