+++ src/network.cpp
@@ -13,8 +13,8 @@ void Network::init_network(unsigned int *inputs, unsigned int *neurons,
   }
 }
 
-void Network::propagate_network(const double *input) {
-  std::memcpy(layers[0]->input, input, layers[0]->count_input * sizeof(double));
+void Network::propagate_network(const float *input) {
+  std::memcpy(layers[0]->input, input, layers[0]->count_input * sizeof(float));
 
   for (unsigned int l = 0; l < count_layers; l++) {
     layers[l]->propagate_layer();
@@ -26,12 +26,12 @@ void Network::propagate_network(const double *input) {
   }
 }
 
-double Network::train_network(const double *input, const double *awaited_output,
-                              const double learning_rate, double momentum) {
+float Network::train_network(const float *input, const float *awaited_output,
+                              const float learning_rate, float momentum) {
 
   propagate_network(input);
-  double total_error = 0;
-  double out, delta, dw;
+  float total_error = 0;
+  float out, delta, dw;
 
   Layer *output_layer = layers[count_layers - 1];
